Bukkit/../PlayerEvents - BlockDispenseEvent : fix cancelation of all DispenseEvents but Water and Lava
Calculating the target position of dispensed item doesnt work as expected, but also seems not needed here.
Dropping items on roads is not a problem, dispensing arrows or other dispensable items will be permitted by missing buildpermissions on roads and nearby plots, so just focusing on Water and Lava dispensed on roads and cancel only those events, but ignore all other.